Power usage bug work-around for tuna SMC.
This commit implements the work-around for the power usage bug in tuna SMC
PA and enables SMC-related kernel options. Note that in order to take
advantage of this code there are also some userspace changes that are
required, e.g. enabling SMC PA loading and using tuna keymaster version
that is ported to Keymaster HAL 0.

Conflicts:
	arch/arm/configs/tuna_defconfig

Change-Id: I1e698941855bc91b29dd759c983470b8390ee1a4